% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcovwrapper.R
\name{d_covR}
\alias{d_covR}
\title{d_covR}
\usage{
d_covR(X, Y, N)
}
\arguments{
\item{X}{- vector of random samples fron the distribution of X}

\item{Y}{- vector of random samples from the distribution of Y}

\item{N}{- scalar positive integer valued parameter to be supplied by the user, must be less than sqrt(length(x))}
}
\value{
distance covariance between X and Y via a fast algorithm
}
\description{
d_covR
}
