% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalkernel.R
\name{optkern}
\alias{optkern}
\title{Title}
\usage{
optkern(X, Y, N, alfa = 0.05)
}
\arguments{
\item{X}{an nxd data matrix containing n vectors from the distribution of X, each of length d}

\item{Y}{an nxd data matrix containing n vectors from the distribution of Y, each of length d}

\item{N}{a parameter to be chosen by the user, ideally close to half of n.}

\item{alfa}{significance level, automatically taken as 0.05}
}
\value{
optimal kernel test statistic, the critical point and the decision whether the test is rejected
}
\description{
Title
}
